Fixed product type

+++ untitled/src/Product.java
@@ -1,10 +1,9 @@
 public class Product {
     private int id;
     private String name;
-    private int price;
+    private double price;
     private String category;
     private int stockQuantity;
-
 
     public int getId() {
         return id;
@@ -22,11 +21,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
